+++ src/config.js
@@ -1,0 +1,7 @@
+// src/config.js
+
+/**
+ * API의 기본 URL을 정의하는 전역 설정 파일입니다.
+ * 서버 주소가 변경될 경우 이 파일의 값만 수정하면 됩니다.
+ */
+export const API_BASE_URL = 'https://www.jungjiyu.com';
