+++ src/pages/Auth/CompleteUserSignUpForm.jsx
@@ -1,7 +1,7 @@
 import React, { useState } from 'react';
 import { useNavigate } from 'react-router-dom';
 // import { completeUserSignUp } from '../api/authApi'; // 삭제
-import axiosInstance from '../api/axiosInstance'; // axiosInstance 임포트
+import axiosInstance from '../../api/auth/axiosInstance'; // axiosInstance 임포트
 import './Auth.css';
 
 export default function CompleteUserSignUpForm() {
