+++ src/pages/Auth/SignUpStep2Page.jsx
